notice-reminders/pyproject.toml
[project]
name = "notice-reminders"
version = "0.1.0"
description = "MOOC Course Search & Announcement Reminders"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.1",
    "tortoise-orm>=0.20.0",
    "aerich>=0.7.2",
    "pydantic-settings>=2.2.1",
    "email-validator>=2.1.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.3",
    "python-multipart>=0.0.9",
    "typing-extensions>=4.8.0",
]

[project.scripts]
notice-reminders = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.pyright]
reportIncompatibleVariableOverride = false
include = ["."]
exclude = [".venv", "**/__pycache__"]
venvPath = "."
venv = ".venv"
